Decompose Codex cost scanner
